Move UTF-8 tests to separate file
