+++ test/with_api_v2/acceptance/api_test.go
@@ -15,7 +15,6 @@ package test
 
 import (
 	"fmt"
-	"strings"
 	"testing"
 	"time"
 
@@ -29,7 +28,7 @@ import (
 	a "github.com/prometheus/alertmanager/test/with_api_v2"
 )
 
-func TestAddClassicAlerts(t *testing.T) {
+func TestAddAlerts(t *testing.T) {
 	t.Parallel()
 
 	conf := `
@@ -58,7 +57,6 @@ receivers:
 
 	am := amc.Members()[0]
 
-	// can create an alert with classic labels
 	now := time.Now()
 	pa := &models.PostableAlert{
 		StartsAt: strfmt.DateTime(now),
@@ -77,81 +75,6 @@ receivers:
 
 	_, err := am.Client().Alert.PostAlerts(alertParams)
 	require.NoError(t, err)
-
-	// cannot create an alert with UTF-8 labels
-	now = time.Now()
-	pa = &models.PostableAlert{
-		StartsAt: strfmt.DateTime(now),
-		EndsAt:   strfmt.DateTime(now.Add(5 * time.Minute)),
-		Alert: models.Alert{
-			Labels: models.LabelSet{
-				"a":                "a",
-				"00":               "b",
-				"Σ":                "c",
-				"\xf0\x9f\x99\x82": "dΘ",
-			},
-		},
-	}
-	alertParams = alert.NewPostAlertsParams()
-	alertParams.Alerts = models.PostableAlerts{pa}
-
-	_, err = am.Client().Alert.PostAlerts(alertParams)
-	require.NotNil(t, err)
-	require.True(t, strings.Contains(err.Error(), "invalid label set"))
-}
-
-func TestAddUTF8Alerts(t *testing.T) {
-	t.Parallel()
-
-	conf := `
-route:
-  receiver: "default"
-  group_by: []
-  group_wait:      1s
-  group_interval:  10m
-  repeat_interval: 1h
-receivers:
-- name: "default"
-  webhook_configs:
-  - url: 'http://%s'
-`
-
-	at := a.NewAcceptanceTest(t, &a.AcceptanceOpts{
-		Tolerance: 1 * time.Second,
-	})
-	co := at.Collector("webhook")
-	wh := a.NewWebhook(t, co)
-
-	amc := at.AlertmanagerCluster(fmt.Sprintf(conf, wh.Address()), 1)
-	require.NoError(t, amc.Start())
-	defer amc.Terminate()
-
-	am := amc.Members()[0]
-
-	now := time.Now()
-	labels := models.LabelSet{
-		"a":                "a",
-		"00":               "b",
-		"Σ":                "c",
-		"\xf0\x9f\x99\x82": "dΘ",
-	}
-	pa := &models.PostableAlert{
-		StartsAt: strfmt.DateTime(now),
-		EndsAt:   strfmt.DateTime(now.Add(5 * time.Minute)),
-		Alert:    models.Alert{Labels: labels},
-	}
-	alertParams := alert.NewPostAlertsParams()
-	alertParams.Alerts = models.PostableAlerts{pa}
-
-	_, err := am.Client().Alert.PostAlerts(alertParams)
-	require.NoError(t, err)
-
-	resp, err := am.Client().Alert.GetAlerts(nil)
-	require.NoError(t, err)
-	require.Len(t, resp.Payload, 1)
-
-	alert := resp.Payload[0]
-	require.Equal(t, labels, alert.Labels)
 }
 
 // TestAlertGetReturnsCurrentStatus checks that querying the API returns the
